refactor permission helpers/hooks to correct folders
